+++ lib/src/persistor/stubs/stub_sqlite_persistor.dart
@@ -1,0 +1,17 @@
+import 'package:loon/src/persistor/stubs/stub_persistor.dart';
+
+class SqlitePersistor extends StubPersistor {
+  SqlitePersistor({
+    super.encrypter,
+    super.onClear,
+    super.onClearAll,
+    super.onHydrate,
+    super.onPersist,
+    super.onSync,
+    super.persistenceThrottle,
+    super.settings,
+    this.useFfi = false,
+  });
+
+  final bool useFfi;
+}